Dockerfile: update rust image
The exisitng image is based on debian buster.  It's 2025, and the apt
repos are no longer available at their original location.  Hence all the
apt-get commands are failing.

I'm sure there are archive mirrors somewhere, but rather than continue
to live in the past, update to a new rust image